fixup! Add configurable resource quotas for the service and spawned jobs
