+++ pkg/k8sutils/connect.go
@@ -18,7 +18,7 @@ type k8sImpl struct {
 // K8s is used to interact with kubernetes jobs
 type K8s interface {
 	ConnectToCluster() error
-	CreateK8sJob(jobName string, action *config.Action, task config.Task, eventData *keptnv2.EventData, jobSettings config.JobSettings, jsonEventData interface{}) error
+	CreateK8sJob(jobName string, action *config.Action, task config.Task, eventData *keptnv2.EventData, jobSettings JobSettings, jsonEventData interface{}) error
 	DeleteK8sJob(jobName string) error
 	GetLogsOfPod(jobName string) (string, error)
 }
